refactor: move cancel-run from infra to zero layer (#8365)
Move cancelRun() and CancelRunResult from infra/run/run-service.ts
to zero/zero-run-cancel.ts. Cancel logic involves queue entry deletion
which is a queue/business concern owned by the zero layer.

- Create zero-run-cancel.ts with cancelRun and exported CancelRunResult
- Update both cancel API routes to import cancelRun from zero layer
- Keep dispatchCancelSideEffects in infra (Ably/callback concerns)
- Remove unused agentRunQueue import from run-service.ts

Pure code relocation — no behavioral changes.
